#535 dynamicExtent vector layer and filtering issues (#536)
